requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
Automat==22.10.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
boto3==1.26.49
botocore==1.29.49
bs4==0.0.1
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
cssselect==1.2.0
Django==4.1.4
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-filter==22.1
django-model-utils==4.3.1
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.2.2
dnspython==2.2.1
drf-yasg==1.21.4
drf-yasg-json-api==0.9.1
humanize==4.4.0
psycopg2-binary==2.9.5
PyJWT==2.4.0
pymongo==4.3.3
PyNaCl==1.5.0
pyOpenSSL==22.1.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
queuelib==1.6.2
redis==4.4.1
redis-server==6.0.9
requests==2.28.1
requests-file==1.5.1
ruamel.yaml==0.17.21
urllib3==1.26.13
vine==5.0.0
w3lib==2.1.1
wcwidth==0.2.5
whitenoise==6.3.0
zope.interface==5.5.2
python-dotenv==0.21.0
BeautifulSoup4==4.11.1
channels==4.0.0
channels-redis==4.0.0
h2==4.1.0
gunicorn==20.1.0
uvicorn==0.20.0
pytest==7.2.1
h2==4.1.0
daphne==4.0.0
psycopg2-binary==2.9.5
redis-server==6.0.9
requests==2.28.1
BeautifulSoup4==4.11.1
google-auth==1.21.2
django-storages==1.13.2